Refactor PT1: switch statement
Use switch statement for running all of the opcodes. This gave me a 50ms speed improvement when running the tests

+++ LibDotNetParser/CILApi/IL/ILInstruction.cs
@@ -16,7 +16,7 @@ namespace LibDotNetParser.CILApi
         /// <summary>
         /// The opcode
         /// </summary>
-        public int OpCode { get; set; }
+        public ushort OpCode { get; set; }
         /// <summary>
         /// The operand
         /// </summary>
